backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.3.0
aiohttp>=3.9.0
redis>=5.0.1 
requests>=2.31.0
html2text>=2020.1.16
apscheduler>=3.10.0
lxml>=4.9.3
httpx>=0.26.0
python-multipart>=0.0.9
yfinance>=0.2.36
google-auth>=2.23.0
PyJWT>=2.8.0
